Completed contents outline up to current work, fixed font problems
The thesis now gives a skeleton argument for the available data - flesh now needs to be added to to be added to the bones.
The problem with small caps Æ being replaced with æ has been fixed thanks to help from Stackexchange, though it has been suggested that changing to EBGaramond would be better.